+++ functions.ts
@@ -1,3 +1,4 @@
 //TODO: make the type more user friendly, currently it requires Uint8Array
-export { validateRows } from "./src/validators/row-validator/rowValidator";
-export { validateColumns } from "./src/validators/column-validator/columnValidator";
+export { validateRows } from "./src/validators/row-validator/validateRows";
+export { validateColumns } from "./src/validators/column-validator/validateColumns";
+export { validateThermos } from "./src/validators/thermo-validator/validateThermos";
