+++ src/validators/thermo-validator/validateThermos.ts
